test(graph): adapt and colocate prune tests
Use current main error/Neighbor APIs, record the existing non-finite-alpha limitation without changing behavior, and keep private tests beside their implementation.
